requirements_pi.txt
jax[cpu]
flax
numpy
pyserial